refactor: Modularize the agent - break it into smaller pieces that allow to get a better clarity about what files are scaffolds and what are dynamic
- Introduced WeatherData struct to represent weather information.
- Created FetchWeatherParams struct for fetching weather parameters.
- Implemented WeatherService interface and MockWeatherService for generating mock weather data.
- Added methods to fetch weather and get available conditions.
- Implemented WeatherToolHandler to handle weather-related tool calls.
- Added utility function stringPtr for string pointer creation.
