+++ examples/docker-compose/a2a/weather-agent/utils.go
@@ -1,0 +1,6 @@
+package main
+
+// stringPtr returns a pointer to the given string
+func stringPtr(s string) *string {
+	return &s
+}
